+++ src/Datasets/Labeled.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\Datasets;
 
+use Rubix\ML\Transformers\Transformer;
 use InvalidArgumentException;
 use RuntimeException;
 
@@ -108,15 +109,9 @@ class Labeled extends DataFrame implements Dataset
             $labels = array_values($labels);
 
             foreach ($labels as &$label) {
-                if (!is_string($label) and !is_numeric($label)) {
-                    throw new InvalidArgumentException('Label must be a string'
-                        . ' or numeric type, ' . gettype($label) . ' found.');
-                }
-
-                if (is_string($label) and is_numeric($label)) {
-                    $label = (int) $label == $label
-                        ? (int) $label
-                        : (float) $label;
+                if (!is_string($label) and !is_numeric($label) and !is_array($label)) {
+                    throw new InvalidArgumentException('Label must be a string,'
+                        . ' numeric, or array type, ' . gettype($label) . ' found.');
                 }
             }
         }
@@ -173,6 +168,19 @@ class Labeled extends DataFrame implements Dataset
     public function possibleOutcomes() : array
     {
         return array_values(array_unique($this->labels));
+    }
+
+    /**
+     * Apply a transformation to the dataset and return for chaining.
+     *
+     * @param  \Rubix\ML\Transformers\Transformer  $transformer
+     * @return self
+     */
+    public function apply(Transformer $transformer) : self
+    {
+        $transformer->transform($this->samples, $this->labels);
+
+        return $this;
     }
 
     /**