+++ src/Transformers/L2Normalizer.php
@@ -17,12 +17,13 @@ use InvalidArgumentException;
 class L2Normalizer implements Transformer
 {
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         foreach ($samples as &$sample) {
             $norm = 0.;