+++ src/Transformers/LambdaFunction.php
@@ -32,13 +32,14 @@ class LambdaFunction implements Transformer
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
-        $samples = call_user_func($this->lambda, $samples);
+        list($samples, $labels) = call_user_func($this->lambda, $samples, $labels);
     }
 }