+++ src/Transformers/LinearDiscriminantAnalysis.php
@@ -178,13 +178,14 @@ class LinearDiscriminantAnalysis implements Transformer, Stateful
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @throws \RuntimeException
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         if (is_null($this->eigenvectors)) {
             throw new RuntimeException('Transformer has not been fitted.');