+++ src/Transformers/MinMaxNormalizer.php
@@ -156,13 +156,14 @@ class MinMaxNormalizer implements Transformer, Elastic
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @throws \RuntimeException
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         if (is_null($this->mins) or is_null($this->scales)) {
             throw new RuntimeException('Transformer has not been fitted.');