+++ src/Transformers/PolynomialExpander.php
@@ -42,12 +42,13 @@ class PolynomialExpander implements Transformer
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         $columns = count(current($samples));