+++ src/Transformers/RobustStandardizer.php
@@ -95,13 +95,14 @@ class RobustStandardizer implements Transformer, Stateful
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @throws \RuntimeException
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         if (is_null($this->medians) or is_null($this->mads)) {
             throw new RuntimeException('Transformer has not been fitted.');