+++ src/Transformers/TfIdfTransformer.php
@@ -117,13 +117,14 @@ class TfIdfTransformer implements Transformer, Elastic
     }
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @throws \RuntimeException
      * @return void
      */
-    public function transform(array &$samples) : void
+    public function transform(array &$samples, ?array &$labels = null) : void
     {
         if (is_null($this->idfs)) {
             throw new RuntimeException('Transformer has not been fitted.');