+++ src/Transformers/Transformer.php
@@ -9,10 +9,11 @@ interface Transformer
     const EPSILON = 1e-8;
 
     /**
-     * Transform the sample matrix.
+     * Transform the dataset in place.
      *
      * @param  array  $samples
+     * @param  array|null  $labels
      * @return void
      */
-    public function transform(array &$samples) : void;
+    public function transform(array &$samples, ?array &$labels = null) : void;
 }